move getApp(file, path) to router mixin
